lstiter

+++ src/ft_substr.c
@@ -4,10 +4,10 @@
 
 char *ft_substr(char const *s1, unsigned int start, size_t len)
 {
-    int i;
+    size_t i;
     char *new_string;
 
-    new_string = malloc((sizeof (char) * len) + 1);
+    new_string = malloc((sizeof(char) * len) + 1);
     if (new_string == NULL)
         return (0);
     i = 0;
